Improve check for whether canvas is visible
As seen with https://stackoverflow.com/questions/57676500 scaling the
canvas causes problems for glslCanvas.  This is because checking for
whether the canvas is visible mixes values from the bounding client rect
(which is affected by scaling), with the canvas element (which is not).

Use the bounding client rect to determine the visibility of the canvas.